// lib/models/IStoreResponse.ts
export default interface IStoreResponse {
   storeid: number;
   address1: string;
   address2: string;
   city: string;
   lat: number;
   lng: number;
   distance: number;
   province: string;
   postal: string;
   phone: string;
   timcard: boolean;
   instoreseating: boolean;
   hrs24: boolean;
   coldstone: boolean;
   wifi: boolean;
   servicedisruptionflag: boolean;
   drivethru: boolean;
   doubledrivethru: boolean;
   catering: boolean;
   selfservice: boolean;
   coldstonesoftswirl: string;
   kurig: boolean;
   tassimo: boolean;
   nfc: boolean;
   barcode: boolean;
   id: number;
}